+++ DesignerMacros/src/common/attributeWithMapPath.ts
@@ -1,0 +1,8 @@
+interface IAttributeWithMapPath {
+    id: string,
+    name: string,
+    typeId: string,
+    mapPath: string[],
+    isNullable: boolean,
+    isCollection: boolean
+};