+++ DesignerMacros/src/services-crud/create-crud/create-crud.ts
@@ -2,6 +2,7 @@
 /// <reference path="../../common/domainHelper.ts" />
 /// <reference path="../../common/crudHelper.ts" />
 /// <reference path="../../common/servicesHelper.ts" />
+/// <reference path="../../common/elementManager.ts" />
 /// <reference path="../../services-cqrs-crud/_common/onMapDto.ts" />
 
 const privateSettersOnly = application.getSettings("c4d1e35c-7c0d-4926-afe0-18f17563ce17")?.getField("0cf704e1-9a61-499a-bb91-b20717e334f5")?.value == "true";